match.c: match_word(): define static "corner case regexes" data
This new structure, mword_regexes, contains four members, and two static
structures are defined for each of the separators. The members of this
structure are:

- re_double_sep: anchored regex matching two separators;
- re_separator: regex matching the separator;
- re_begin_full: regex matching the separator, anchored at the beginning;
- re_end_full: regex matching the separator, andchored at the end.

The re_double_sep member is only ever used if the glob is meant to match only
the separator; others are used when generating data for amglob_to_regex().

There are several benefits:

* no need to compute the beginning and end anchors anymore, they are statically
  defined and only depend on which separator is used;
* in the "separator only" case, we can calculate the match result early, no need
  to allocate space for the regex string anymore;
* the regex local variable can be pushed down, and one more local variable (dst)
  eliminated.