Move column spec validation to conffile.c, get rid of columnar.[ch]
Prior to this, conffile.c was not checking the column spec validity. This task
was delegated to amreport and amcheck respectively. For amcheck in particular,
it meant that it held for itself two entire files, common-src/columnar.[ch].

Now that it is up to conffile.c to validate these, the check in amcheck.c can be
removed and the two columnar.* files deleted.