[Coccinelle] Add macro to use for spatch's -macro_file_builtins option
This allows spatch to parse 99+% of the code correctly. It still chokes on
client-src/selfcheck.c: this is an spatch bug.

To use spatch:

* be at the root of the Amanda source tree;
* type:

for dir in *-src; do
	spatch -sp_file /path/to/file.cocci \
		-macro_file_builtins coccinelle/coccinelle-builtins.h
		-in_place -dir $dir;
done

If you want to check whether spatch is able to parse a file, use:

spatch -parse_c -macro_file_builtins coccinelle/coccinelle-builtins.h \
	/path/to/file.c