[allocapi] common-src/util.c: remove usage of vstrextend()
The debug_executing() function builds a shell-quoted string of a command line
(except of the command itself), with all arguments passed in a GPtrArray.
Unfortunately, this means we cannot use g_ptr_array_free(..., FALSE) as the
caller expects its array to be intact on exit. We have to peek into the
structure itself. Do that, and put the resulting quoted command line in a
GString.