Change prototype for validate_optionstr(), and fix two memory leaks
validate_optionstr() returned a GPtrArray of statically allocated strings to the
caller, it was then up to the caller to verify whether the array was non empty
by peeking into it, which defeated the purpose of using a GPtrArray in the first
place: you just shouldn't do that.

What's more, it was freed with g_ptr_array_free(..., TRUE), which attempted to
free individual elements. GLib developers should be credited for having handled
this case correctly, but still.

Change the function to return a gchar ** of dynamically allocated strings
instead (NULL if no errors), and change all four callsites. What's more, this
fixes two callsites, which would leak memory because they didn't free the array
when done (server-src/driver.c in read_schedule(), and server-src/planner.c in
getsize()).

[maybe the fact that the latter forgot to free explains why installchecks passed
for it: when I started to free correctly, glibc warned me about an invalid freed
pointer]