server-src/diskfile.c: fix potential SIGSEGV
In both the optionstr() and xml_optionstr() functions, the "their_features"
member was assigned to _before_ the asserts on dp and dp->host were done. This
could result in a potential SIGSEGV should any of these were indeed NULL.

Postpone the assignment to only after the asserts were done. The program will
abort anyway, but at least there will be logs.

Also, in both of these functions:

* replace assert() with g_assert();
* remove unused local variable "qdpname".