server-src/planner.c: build the initial request string with a GString
Prior to this, it was built using g_strjoin(), with the added malus that the
number to be sorted in a string was strprintf()ed into an on-stack buffer.

Change that by using a GString and append to it on demand instead, which has the
additional benefit of making this part of the code much cleaner.