common-src/conffile.c: add execute_on_to_string() helper
The CONFTYPE_EXECUTE_ON property type has a lot of values, and two call sites
deal with these values in order to build a string representation of it.

As it is quite easy to miss a site when adding a flag, add a helper function to
return the resulting string to the caller.