[allocapi] client-src/sendsize.c: remove usage of vstrextend(), 1/2 [FIXME]
One usage of vstrextend() was in generic_calc_estimates(), to build a string
version of a command line stored in a GPtrArray for display, by accessing its
->ptr member directly (don't do that). The GPtrArray was then extended to add
further arguments.

Use g_ptr_array_free(..., FALSE) to gather the list of args at the appropriate
point to dump the string, then reallocate the array with the same strings in it,
and expand it before using g_ptr_array_free(..., FALSE) again to pass the
arguments to pipespawnv().  This avoids accessing the struct members.

[FIXME: why stop in the middle? The added arguments afterwards are also
logged...]

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4112 a8d146d6-cc15-0410-8900-af154a0219e0