[allocapi] client-src/sendsize.c: remove usage of vstrextend(), 2/2 [FIXME]
The second usage of vstrextend() was in getsize_application_api(), to build a
modified string version of a command line stored in a GPtrArray for display (the
only difference is, we log the full command name, but leave up the short command
name in what is passed to execve()). Use g_ptr_array_free(..., FALSE) and
g_strjoinv() instead. This avoids accessing the GPtrArray private members, which
you just should not do.

[FIXME: why not pass the full command to execve()?]

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4113 a8d146d6-cc15-0410-8900-af154a0219e0