client-src/calcsize.c: simplify, and accelerate, rounded block size calculation
The estimated block size, as default or for tar, was calculated as such:

----
 #define ST_BLOCKS(s)                                                          \
            (((((off_t)(s).st_blocks * (off_t)512) <= (s).st_size)) ?          \
              ((off_t)(s).st_blocks + (off_t)1) :                              \
              ((s).st_size / (off_t)512 +                                      \
                (off_t)((((s).st_size % (off_t)512) != (off_t)0) ?             \
                (off_t)1 : (off_t)0)))
----

which calculates the upper round of the number blocks to the nearest blocksize.
It was then used for dump as:

----
 (ST_BLOCKS(s) + 1) / 2
----

as a block is 1k and not 512 bytes.

But here, block sizes are always powers of 2. And, for a given power of 2, there
is a much simpler way to calculate the upper round. For any integer x and n,
where the blocksize is 1 << n, this rounded value can be calculated as:

(x + ((1 << n) - 1)) & ~((1 <<n) - 1)

But here we shift by the block's power of two itself, which avoids even the &.
The calculation therefore becomes:

(x + ((1 << n) -1)) >> n

Take advantage of that.