[allocapi] server-src/diskfile.c: remove usage of vstrextend(), 2/7
The xml_optionstr() function built an XML fragment out of the options of a DLE.
This was done using a flurry of local variables, all carefully crafted to end
with a \n, joined at the end using g_strjoin(NULL, ...).

Change that completely: instead, use a single GPtrArray to collect all options,
this time without a \n at the end, and use the classical g_ptr_array_free(),
g_strjoinv() and g_strfreev() "magic" to build the final result.

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4140 a8d146d6-cc15-0410-8900-af154a0219e0