match.c: match_word(): nglob variable pushdown
This variable was allocated a full copy of the glob argument, which was only of
use if the glob was not of the "separator only" type.

Push it down to the "non separator only" case (along with its allocation and
freeing), which also gets rid of the len variable. Also rename it to glob_copy
along the way, since nglob didn't really mean anything.

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4165 a8d146d6-cc15-0410-8900-af154a0219e0