match.c: match_word(): initialize static data pointers earlier
Define a boolean telling whether the separator is NOT '/' (which is probably the
most common case). By default, the "regexes" (a struct mword_regexes *) and
"table" (a struct subst_table *) variables are set to the '/' case, and only set
to the "not '/'" (ie, '.') case if need be.

This further improves the code readability.

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4168 a8d146d6-cc15-0410-8900-af154a0219e0