match.c: regex compile caching
Instead of recompiling a regex each time, implement a regex cache in the shape
of a hash table, with the regex string as a key and the compiled regex buffer as
a value.

While this has the drawback that all regex buffers are dynamically allocated,
and regex strings duplicated, this has the potentially huge benefit that when a
regex string is used many times, the regex compile phase is skipped altogether -
instead, the compiled regex buffer can be handed back immediately.

In order to take advantage of this, create a new function
get_regex_from_cache(), which only allocates, compiles and stores the regex if
if it isn't found in the cache. This function uses do_regex_compile() to
generate the compiled regex buffer.

Use this function in all call sites previously calling upon do_regex_compile(),
except validate_regex(), which never matches and for which storing the regex
compile result is not required.

Conflicts:

	common-src/match.c

git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@4170 a8d146d6-cc15-0410-8900-af154a0219e0