server-src/amcheck/c: fix test in check_tapefile()
In this function, stat(2) was called to a file. The next test, if stat(2)
returned 0, was then to check whether the file was a regular file.

The test after that was access(..., F_OK). But that is redundant: if this were
not the case, stat(2) would have returned -1 anyway. This test is therefore
meaningless.

Assume here that R_OK was meant instead, ie that the file is readable by the
user running amcheck. Correct wording of the following error message if
access(2) does not return 0.