Drop use of --device in inner_restore.
Although adapted from amraw's example, for inner_restore to refer to
--device (in determining the directory to restore into) is *not* like
what other standard amanda applications do; they restore right into
amrecover's current directory (or the one given with --directory, if any),
as you would expect when --device held a filesystem and the pathnames
backed up were relative to the root of that filesystem. To uphold the
principle of least astonishment, make this application restore the same way.

Add manpage.