Don't preserve times in capture_rsync_state.
Prior states of the opaque tree can be stored efficiently using
--link-dest, so only changed files are added in a new state while
unchanged ones are made links into the state it is based on.
That way (assuming many files are untouched between backups),
state for n levels does not require n times the space of the tree
being backed up, and often a multiple only slightly greater than 1.

However, commonly the timestamps will be meaningless in the tree
being backed up. That happens if a package provides a tool for
generating a consistent snapshot of its files (svnadmin hotcopy
for subversion, db2bak for an LDAP server, etc.) and if that tool
doesn't preserve timestamps in the tree copy that it writes.

In that case, every tree that amopaquetree is asked to back up
will have new timestamps on every file, even those that (in the
live application) have been untouched since the last backup.
If capture_rsync_state is trying to preserve timestamps, it will
fail to find any files it can link, and the state storage goes
back to requiring n times the tree size for n levels of state.

Therefore, avoid preserving timestamps in capture_rsync_state.

No corresponding change to generate_rsync_batch; it still
preserves times. Accordingly, each file in a restored tree will
end up stamped with the time of the latest dump it came from,
which is not unreasonable, and better than giving everything the
arbitrary new timestamp the application's snapshot tool may have
slapped on it.