Allow mount options for amlvmsnapshot.
This turns out to be necessary for XFS filesystems, where
a newly-created snapshot can't be mounted without the nouuid
option (because, unsurprisingly, it has the same uuid as its
origin), or without the norecovery option. That last is surprising,
because there is an xfs_freeze operation and, per documentation,
it is automatically called by the dm driver when an LVM snapshot
is created, so the snapshot should be of a quiescent filesystem,
but somehow the process results in a couple snapshot-related
entries in the journal, enough for mount to think recovery is needed.