Add more stages where amlvmsnapshot can be run.
When space permits, it seems ideal to run amlvmsnapshot
only at pre-dle-estimate to create the snapshot, and at
post-dle-backup to free it, thereby backing up exactly
what was estimated. But if there was not much volume-group
space left to allocate to a snapshot, or if Amanda's delay
in gathering all estimates and planning backups is long,
it could be possible for the snapshot to run out of space.
For that case, allow defining a dumptype that runs
amlvmsnapshot four times (pre/post-dle-estimate,
pre/post-dle-backup). In that case, the backup will be done
from a second snapshot taken later, so it won't be exactly
what was estimated (that's why it's called an estimate ;)
but the two snapshots will be shorter-lived and less likely
to exhaust available space.

Allow amlibvirtfsfreeze to run pre-dle-backup too. Add man pages.