Allow amlibvirtfsfreeze without mountpoints.
In RHEL7, the qemu-guest-agent is able to freeze
selected filesystems, specified by their mountpoints
(though when it comes time to unfreeze, no mountpoints
can be specified, all frozen filesystems are unfrozen
at once). However, RHEL6's guest agent is not able to
freeze selected filesystems, only all of them at once.

So, the case of 'freeze' with no mountpoints specified
can't be rejected as an error; it could be necessary
for a RHEL6 VM.