Let application override Amanda localstatedir.
This can be useful if, for example, backing up something that lives
on a VM and the default Amanda::Paths::localstatedir ends up on the
root filesystem, which might be a sparsely-provisioned or copy-on-
write image that you'd rather avoid continually writing into, so
it stays more or less reflective of the OS itself and its updates.

If the VM is also provided another filesystem more appropriate for
continual day-to-day scribbling, this property can move the Amanda
local state to there.