Now with amgrowingfile application.
Allows writing a DLE that refers to a single file that is known
to only grow; can backup and restore incremental levels, which are
simply the tail of the file from where it ended at the prior level.
If the file ever gets rewritten (not just appended), it will be
important to remember to force a level 0 dump.