Add amopaquetree application.
This application can back up and restore, incrementally, a directory
tree of files treated as an indivisible unit (say, something managed
by a DBMS or version-control system, where one is interested in
restoring the state of the whole tree as of a given backup point, but
individual files are of no interest except to the system that manages
them).

What is actually backed up is an rsync 'batch' stream; rsync can
generate an efficiently encoded stream for turning one tree into
another without needing internal knowledge of the data format.
(A level 0 backup is saved as a batch stream to turn an empty directory
into the tree being backed up.)