A script to freeze/thaw filesystems through libvirt.
If a domain running under libvirt is able to respond to the
virsh domfsfreeze/domfsthaw commands (say, it is a qemu domain
with the qemu-guest-agent running within), then this script can
be used to freeze one or more filesystems in the domain and then
thaw them again. The intent is for a DLE for backing up the host
node to mention this script twice (through two different amanda.conf
'script' definitions, one with the freezeorthaw property set to freeze
and 'order' set to a lower number than the amlvmsnapshot script, and
the other with freezeorthaw set to thaw, and 'order' higher than that
of amlvmsnapshot.

The result should be that the running domain has its filesystem(s)
frozen, but only long for the host node to grab an LVM snapshot of
its own filesystem(s), then immediately thawing the domain
filesystem(s) ... all of this in pre-dle-estimate. The snapshot of
the host filesystem can then be used for estimate and backup, and
any guest domain filesystem image file(s) should be in a consistent
state.