Add an am389bak script.
To back up a 389 LDAP directory server instance, this script
can be used on pre-dle-estimate to run db2bak, which copies out
a consistent snapshot of the database files from the server instance
(whose name--the part of its directory name after the slapd- prefix--
has to be specified with the 'instance' property) into the directory
named by the DLE's 'device'. Then amopaquetree is great for dumping
that consistent snapshot.

It sounds easy but permissions complicate matters. If the Amanda user
isn't root and the 389 server runs as a different user, some kind of
setuid 'rundb2bak' wrapper is needed. In fact, it has to be setuid and
setgid AND copy both of those to the real ids before it execs db2bak,
which otherwise complains. That means the wrapper had better live on
a file system that supports ACLs, because with both user and group
having to be 389's, there'd be no other way to make it executable by
the Amanda user but not by everyone.

A couple other annoying things done by db2bak are also best handled
in the setuid wrapper. If the destination directory already exists,
db2bak moves it to the same name with .bak tacked on (there is no
documented option to not do that). By itself that's not so bad, but
if that .bak directory ALSO already exists, db2bak fails. So it has
to be removed every time, most easily in the same setuid wrapper, which
is able to do so.

A tidy way for the backup strategy to work is to have a default, or
inheritable, ACL on the parent directory of the destination (DLE
'device'), so that when db2bak writes the files there (running as
the 389 user), they get ACLs allowing the Amanda user to read them,
so amopaquetree then has no trouble dumping them.

That's another thing db2bak is able to break, by creating its files
and directories with explicit modes disallowing group access (which,
at least in the POSIX ACL world, has the effect of zeroing the ACL's
'mask' entry; the files all inherit the parent ACL giving access to
the Amanda user, but the mask blocks it anyway). There might be some
other way around that, but the setuid wrapper used here also just runs
through the resulting tree fixing the doggoned ACLs.

That wrapper's not included in this commit, out of a sense that it's
probably too specific to this site. For Amanda to really benefit from
easy development of scripts like this one, I think there also needs to
be some kind of generalization of runtar that can allow other things
to be run with privilege, subject to some simple client-host configuration
file limiting what can be run and for what DLEs. Future work....