Let 'local' match the local box however named.
It was not clearly documented what hostname must be used in a DLE for
auth "local" to work, and the test relied on string comparison, which
could report the host as "not local" depending on how the name was
abbreviated or fully qualified, capitalized, aliased, etc.

Instead, allow the DLE to use any hostname that can be confirmed to
refer to this host. As long as getaddrinfo finds at least one address
for it that the local end of a socket can be bound to, it will work.

Clarify that in amanda-auth(7), and also clarify that the "authenticated
peer hostname" will be as reported by gethostname() whenever that
succeeds, or "localhost" if it does not. (This has been true since 3.3.0
but the documentation didn't catch up.)

Discussion: http://marc.info/?t=149202030200002