Sop for OS where getaddrinfo can return garbage.
Open Group Base Specifications for the results from getaddrinfo say
"fields ai_family, ai_socktype, and ai_protocol shall be usable
as the arguments to the socket() function to create a socket
suitable for use with the returned address", but there may be some
operating systems (*cough* solaris 10 *cough*) where the values that
come back will be rejected by socket() as invalid. If so, try again,
passing getaddrinfo an explicit socktype constraint, which seems to
help it return something usable.