Support faster estimates with send -nvP.
Support the OpenZFS 'zfs send -nvP' method of getting an estimated
send size. (Still needs to be tested on a box that supports -nvP.)

Discovered in passing that Amanda::Application::Abstract wasn't
declaring --calcsize as an estimate option if supports_calcsize()
was true ... and fixed a silent, original error in A::A::Abstract
caught by a handy warning from Perl while testing on a different
version.