Call write_local_state from command_backup.
Instead of making inner_backup responsible for calling write_local_state,
have command_backup do that automagically if RECORD is supported and
requested and a $self->{'localstate'} exists.

This is preparation for a future change in which command_backup will get
a confirmation from the server before writing the new state.

Discussion: https://marc.info/?l=amanda-hackers&m=150427714716446

Once it becomes possible that a negative confirmation from the server
prevents write_local_state being called, there should be a
repair_local_state method an application can override to reclaim
any resources that were going to be referred to in the new state,
but would be leaked when that state is not saved.