Admit level-to-snapshot map changes.
Even if there was a consistent mapping of levels to snapshots
as of the last successful backup, creation of new datasets (or
even clones) since then may have messed it up. Try to cope with
the possible mess (in the worst case, by backing up levels, or
even reporting no level is possible but a new level 0).

The requirement (every level maps to the same snapshot name in
every descendant) is too strict. What needs to be confirmed is
that the one level chosen as an increment base satisfies the
requirement.

In passing, found it would be helpful if common_snapshots() worked.