Better encapsulate option handling details.
The original approach to declaring, parsing, and retrieving options
used a 'hacky' convention to store the option value if there was
a validating-code reference occupying its expected spot. (Hacky in
the sense of visible to subclassing code, and creating an arbitrary
restriction on usable names for options.)

Replace with a similar technique using a convention that can't collide
with any usable option name, and a store_option() method that encapsulates
the details so subclassing code does not need to be aware of them.
At option-retrieval time, the details already were encapsulated
within new().

In passing, remove the restriction that option-validating code didn't
work for multivalued options.

Get the current collection of applications and scripts more consistently
using store_option() to set defaults, rather than testing for defined
within new() and applying them there, except for defaults that really do
depend on something only known at instantiation time.

https://marc.info/?l=amanda-hackers&m=150522700909264