Implement 'support' and timestamp property.
As added in upstream at adbcd7f, there is now a timestamp property
(and an implemented support subcommand to advertise it).

Update Amanda::Script::Abstract correspondingly.

Existing scripts were doing various custom property checks in new()
for simplicity; that won't work for 'support' because the properties
are not passed in that case. Therefore, a new method check_properties()
is the place for such checks; it is called by run() before do() in every
case except 'support'.

This is preparatory to a way for scripts to maintain invocation-specific
local state, but that is not in this commit.