Refactor option checking.
Checking can now be done with a sequence of check(condition, message)
(which will report any failed checks without interrupting execution),
followed at the end (in the command_... method) with a single bare
check(), which throws an exception to end execution if any of the
foregoing checks failed.

This should simplify using a single set of check methods both from
command_selfcheck (which simply ought to report as many issues as
possible) and from other commands (which ought to fail if anything
isn't right).